# backend/docs/swagger.yaml
basePath: /
definitions:
...
  dto.AgencyProfileRequest:
    properties:
      company:
        type: string
      email:
        type: string
      logo:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - company
    - email
    - name
    type: object
...
  dto.InvoiceRequest:
    properties:
      amount:
        type: number
      due_date:
        description: 'Format: YYYY-MM-DD'
        type: string
      issue_date:
        description: 'Format: YYYY-MM-DD'
        type: string
      notes:
        type: string
      project_id:
        type: integer
      service_type:
        type: string
      status:
        type: string
    required:
    - amount
    - project_id
    - status
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  dto.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      role:
        type: string
    required:
    - email
    - name
    - password
    - role
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API terpusat untuk sistem manajemen Tim dan Proyek Klien Jalcode.
  title: Jalcode API Documentation
  version: "1.0"
paths:
...
  /api/agency/:
    get:
      description: Mengambil data profil perusahaan, logo, dan kontak utama
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Ambil profil agensi
      tags:
      - Agency
    put:
      consumes:
      - application/json
      description: Memperbarui nama, email, logo, dan detail perusahaan
      parameters:
      - description: Data Agensi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AgencyProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update profil agensi
      tags:
      - Agency
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Melakukan autentikasi menggunakan email dan password untuk mendapatkan
        token JWT
      parameters:
      - description: Data Login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil login & mendapat token
          schema:
            additionalProperties: true
            type: object
      summary: Login pengguna
      tags:
      - Auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Mendaftarkan anggota tim baru ke dalam sistem
      parameters:
      - description: Data Registrasi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Register member baru
      tags:
      - Auth
  /api/auth/update-password:
    put:
      consumes:
      - application/json
      description: Mengganti password pengguna yang sedang login
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update Password
      tags:
      - Auth
...
  /api/invoices/:
    get:
      description: Mengambil daftar tagihan pembayaran
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Ambil semua invoice
      tags:
      - Invoices
    post:
      consumes:
      - application/json
      description: Membuat tagihan baru dengan nomor urut otomatis
      parameters:
      - description: Data Invoice
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.InvoiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Buat invoice baru
      tags:
      - Invoices
  /api/invoices/{id}:
    delete:
      description: Menghapus data invoice yang salah
      parameters:
      - description: ID Invoice
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Hapus tagihan
      tags:
      - Invoices
    put:
      consumes:
      - application/json
      description: Memperbarui nominal atau status tagihan (misal dari Unpaid jadi
        Paid)
      parameters:
      - description: ID Invoice
        in: path
        name: id
        required: true
        type: string
      - description: Data Update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.InvoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update status pembayaran
      tags:
      - Invoices
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"